Move _get_neigbor_pairs back to MACE module now inheritance removed.
